pyproject.toml
[project]
name = "ai-browser"
version = "0.1.0"
description = "A general-purpose AI browser/crawler agent built with LangChain, LangGraph, and Playwright"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.18",
    "langchain-openai>=1.2.1",
    "langchain-anthropic>=1.4.3",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.10",
    "playwright>=1.59",
    "beautifulsoup4>=4.14.3",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_browser"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
